atualização: tipo char foi alterado para string para permitir tamanhos dinamicos e unsignedSmallInteger foi alterado para unsignedBigInteger

+++ database/migrations/2026_02_13_220312_create_pagamentos_table.php
@@ -11,8 +11,8 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::create('pagamentos', function (Blueprint $table) {
-            $table->unsignedSmallInteger('id');
-            $table->unsignedSmallInteger('numero_inscricao');
+            $table->unsignedBigInteger('id');
+            $table->unsignedBigInteger('numero_inscricao');
 
             $table->decimal('valor', 5, 2);
             $table->string('mes_de_referencia', 7);
